补充中间件操作类

+++ SpringCloudTemplate/api-admin/src/main/java/com/kcx/api/admin/sys/dao/SysUserDao.java
@@ -33,6 +33,6 @@ public interface SysUserDao extends BaseMapper<SysUserEntity> {
      * @param req
      * @return
      */
-    int userListTotal(ReqUserListVO req);
+    long userListTotal(ReqUserListVO req);
 }
